requirements.txt
streamlit
spotipy
pandas
matplotlib
python-dotenv
